+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Added Base network support to Thorchain plugin.
+
 ## 2.17.1 (2025-01-27)
 
 - changed: (Exolix) Use S-address for Digibyte swaps