+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: Added Base network support to Thorchain plugin.
 - added: Added orderUri to 0xGasless plugin EdgeTxAction metadata.
+- fixed: (SwapKit) UTXO swap memo handling
 - fixed: Fixed 0xGasless plugin orderId.
 
 ## 2.17.1 (2025-01-27)